Move TestingRowRanges to io.trino.parquet.reader
