+++ lib/trino-parquet/src/test/java/io/trino/parquet/reader/TestingRowRanges.java
@@ -11,9 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.parquet.internal.filter2.columnindex;
+package io.trino.parquet.reader;
 
 import org.apache.parquet.internal.column.columnindex.OffsetIndex;
+import org.apache.parquet.internal.filter2.columnindex.RowRanges;
 
 import java.util.stream.IntStream;
