+++ lib/trino-parquet/src/test/java/io/trino/parquet/reader/flat/TestRowRangesIterator.java
@@ -21,8 +21,8 @@ import java.util.Optional;
 import java.util.OptionalLong;
 
 import static io.trino.parquet.reader.FilteredRowRanges.RowRange;
-import static org.apache.parquet.internal.filter2.columnindex.TestingRowRanges.toRowRange;
-import static org.apache.parquet.internal.filter2.columnindex.TestingRowRanges.toRowRanges;
+import static io.trino.parquet.reader.TestingRowRanges.toRowRange;
+import static io.trino.parquet.reader.TestingRowRanges.toRowRanges;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
